[feat]: [PL-24523]: LDAP Password Secret Reference, Enum Introduction for AuthType. (#32939)
* [feat]: [PL-24523]: LDAP Password Secret Reference Support

* [feat]: [PL-24523]: LDAP Password Secret Reference Support

* [feat]: [PL-24523]: Behind FF LDAP_SECRET_AUTH and Added UTs;

* [feat]: [PL-24523]: Removed unused code

* [feat]: [PL-24523]: LDAP Password Secret Reference

* [feat]: [PL-24523]: LDAP Password Secret Reference, Handled Error cases when FF is not turned on

* [feat]: [PL-24523]: LDAP Password Secret Reference, Enum Introduction for AuthType.

* [feat]: [PL-24523]: LDAP Password Secret Reference, UT fixes

* [feat]: [PL-24523]: dependency hash

* [feat]: [PL-24523]: LDAP Password Secret Reference, UT fixes

+++ 400-rest/src/main/java/software/wings/beans/sso/LdapSettings.java
@@ -109,13 +109,14 @@ public class LdapSettings extends SSOSettings implements ExecutionCapabilityDema
 
   public EncryptedDataDetail getEncryptedDataDetails(SecretManager secretManager) {
     return secretManager
-        .encryptedDataDetails(
-            accountId, LdapConstants.BIND_PASSWORD_KEY, connectionSettings.getEncryptedBindPassword(), null)
+        .encryptedDataDetails(accountId, LdapConstants.BIND_PASSWORD_KEY, connectionSettings.fetchPassword(), null)
         .get();
   }
 
-  public void encryptFields(SecretManager secretManager) {
-    if (!connectionSettings.getBindPassword().equals(LdapConstants.MASKED_STRING)) {
+  public void encryptLdapInlineSecret(SecretManager secretManager) {
+    if (isNotEmpty(connectionSettings.getBindPassword())
+        && !connectionSettings.getBindPassword().equals(LdapConstants.MASKED_STRING)) {
+      connectionSettings.setPasswordType(LdapAuthType.INLINE_SECRET);
       String oldEncryptedBindPassword = connectionSettings.getEncryptedBindPassword();
       if (isNotEmpty(oldEncryptedBindPassword)) {
         secretManager.deleteSecret(accountId, oldEncryptedBindPassword, new HashMap<>(), false);
@@ -129,14 +130,20 @@ public class LdapSettings extends SSOSettings implements ExecutionCapabilityDema
       String encryptedBindPassword = secretManager.saveSecretText(accountId, secretText, false);
       connectionSettings.setEncryptedBindPassword(encryptedBindPassword);
       connectionSettings.setBindPassword(LdapConstants.MASKED_STRING);
+      connectionSettings.setEncryptedBindSecret(null);
+    } else {
+      connectionSettings.setBindPassword(LdapConstants.MASKED_STRING);
     }
   }
 
   public void decryptFields(
       @NotNull EncryptedDataDetail encryptedDataDetail, @NotNull EncryptionService encryptionService) {
-    if (connectionSettings.getBindPassword().equals(LdapConstants.MASKED_STRING)) {
+    if (connectionSettings.getPasswordType().equals(LdapAuthType.INLINE_SECRET)) {
       String bindPassword = new String(encryptionService.getDecryptedValue(encryptedDataDetail, false));
       connectionSettings.setBindPassword(bindPassword);
+    } else {
+      String bindSecret = new String(encryptionService.getDecryptedValue(encryptedDataDetail, false));
+      connectionSettings.setBindSecret(bindSecret.toCharArray());
     }
   }
